Toggle and reset the actor class to force refresh instead of hiding and showing.
Hiding and showing the actor to force a refresh for the status area can
cause problems with some other extensions that put hidden icons there.
(for example GSConnect).
Instead, force GS 3.38 to re-render the actors by changing their class
to a dummy value, and restoring the original class.
This method doesn't need to wait for GS to reenter the idel loop either.